refactor: clean up utils tests
