+++ src/utils/util.test.ts
@@ -7,12 +7,11 @@ import { config } from '../config.js';
 import { createAuth } from './auth.js';
 import { createHtml } from './html.js';
 import { createDate } from './date.js';
-import { createMail } from './mail.js';
 import { createRequest } from './request.js';
 import { db } from '../tests/test-setup.js';
 import { createValidation } from './validation.js';
 import type { ApiKeyPayload, BookmarkToExport } from '../type.js';
-import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from 'vite-plus/test';
+import { beforeAll, beforeEach, describe, expect, it, vi } from 'vite-plus/test';
 
 let validationUtils: ReturnType<typeof createValidation>;
 let authUtils: ReturnType<typeof createAuth>;
@@ -20,7 +19,6 @@ let utilUtils: ReturnType<typeof createUtil>;
 let htmlUtils: ReturnType<typeof createHtml>;
 let dateUtils: ReturnType<typeof createDate>;
 let requestUtils: ReturnType<typeof createRequest>;
-let mailUtils: ReturnType<typeof createMail>;
 
 beforeAll(async () => {
     const { createBookmarksRepository } = await import('../routes/bookmarks/bookmarks.repository');
@@ -56,10 +54,8 @@ beforeAll(async () => {
     };
 
     utilUtils = createUtil(mockContext);
-    mailUtils = createMail(mockContext);
 
     mockContext.utils.util = utilUtils;
-    mockContext.utils.mail = mailUtils;
 });
 
 describe.concurrent('truncateString', () => {
@@ -150,32 +146,6 @@ describe.concurrent('isUrlLike', () => {
         expect(validationUtils.isUrlLike(' google.com ')).toBeTruthy();
         expect(validationUtils.isUrlLike('a.b')).toBeFalsy();
         expect(validationUtils.isUrlLike('localhost')).toBeFalsy();
-    });
-});
-
-describe.concurrent('normalizeUrl', () => {
-    it('should return valid https URL as-is', () => {
-        expect(utilUtils.normalizeUrl('https://example.com')).toBe('https://example.com');
-    });
-
-    it('should return valid http URL as-is', () => {
-        expect(utilUtils.normalizeUrl('http://example.com')).toBe('http://example.com');
-    });
-
-    it('should prepend https:// when URL has no protocol', () => {
-        expect(utilUtils.normalizeUrl('example.com')).toBe('https://example.com');
-    });
-
-    it('should prepend https:// for www URLs without protocol', () => {
-        expect(utilUtils.normalizeUrl('www.example.com')).toBe('https://www.example.com');
-    });
-
-    it('should not double-prepend for URLs starting with http', () => {
-        expect(utilUtils.normalizeUrl('http://example.com')).toBe('http://example.com');
-    });
-
-    it('should handle empty string', () => {
-        expect(utilUtils.normalizeUrl('')).toBe('https://');
     });
 });
 
@@ -1031,366 +1001,6 @@ describe('formatDateInTimezone', () => {
         // 13:30 UTC = 9:30 AM NYC
         expect(result2.timeString).toBe('9:30 AM');
         expect(result2.timeInputValue).toBe('09:30');
-    });
-});
-
-describe('sendReminderDigestEmail', () => {
-    let sendMailMock: ReturnType<typeof vi.fn>;
-    let testMailUtils: ReturnType<typeof createMail>;
-
-    beforeAll(async () => {
-        const { createSettingsRepository } = await import('../routes/admin/settings.repository');
-
-        sendMailMock = vi.fn().mockResolvedValue({ messageId: 'test-id' });
-
-        const mockNodemailer = {
-            createTransport: vi.fn().mockReturnValue({ sendMail: sendMailMock }),
-        };
-
-        const prodConfig = {
-            ...config,
-            app: { ...config.app, env: 'production' },
-        };
-
-        const mockLogger = {
-            error: vi.fn(),
-            info: vi.fn(),
-            box: vi.fn(),
-            table: vi.fn(),
-            tag: vi.fn().mockReturnThis(),
-        };
-
-        const testContext = {
-            db,
-            config: prodConfig,
-            libs: { ...libs, nodemailer: mockNodemailer },
-            logger: mockLogger,
-            models: { settings: createSettingsRepository({ db, config, libs } as any) },
-        } as any;
-
-        testMailUtils = createMail(testContext);
-    });
-
-    beforeEach(() => {
-        sendMailMock.mockClear();
-    });
-
-    it('should not send email when reminders array is empty', async () => {
-        await testMailUtils.sendReminderDigestEmail({
-            email: 'test@example.com',
-            username: 'TestUser',
-            reminders: [],
-            date: '2025-08-16',
-        });
-
-        expect(sendMailMock).not.toHaveBeenCalled();
-    });
-
-    it('should send email with clickable links in HTML', async () => {
-        const reminders = [
-            {
-                id: 1,
-                title: 'AI Agent Best Practices',
-                url: 'https://forgecode.dev/blog/ai-agent-best-practices/',
-                reminder_type: 'recurring' as const,
-                frequency: 'weekly' as const,
-            },
-            {
-                id: 2,
-                title: 'Meeting with team',
-                reminder_type: 'once' as const,
-            },
-        ];
-
-        await testMailUtils.sendReminderDigestEmail({
-            email: 'test@example.com',
-            username: 'TestUser',
-            reminders,
-            date: '2025-08-16',
-        });
-
-        expect(sendMailMock).toHaveBeenCalledTimes(1);
-        const emailArgs = sendMailMock.mock.calls[0][0];
-
-        expect(emailArgs.to).toBe('test@example.com');
-        expect(emailArgs.subject).toContain('Reminders');
-        expect(emailArgs.subject).toContain('Saturday, August 16, 2025');
-        expect(emailArgs.html).toContain('Hello TestUser');
-        expect(emailArgs.html).toContain(
-            '<a href="https://forgecode.dev/blog/ai-agent-best-practices/">AI Agent Best Practices</a>',
-        );
-        expect(emailArgs.html).toContain('<li>Meeting with team</li>');
-    });
-
-    it('should show "weekly reminders" when all reminders are weekly recurring', async () => {
-        const reminders = [
-            {
-                id: 1,
-                title: 'Weekly Report',
-                reminder_type: 'recurring' as const,
-                frequency: 'weekly' as const,
-            },
-            {
-                id: 2,
-                title: 'Weekly Review',
-                reminder_type: 'recurring' as const,
-                frequency: 'weekly' as const,
-            },
-        ];
-
-        await testMailUtils.sendReminderDigestEmail({
-            email: 'test@example.com',
-            username: 'TestUser',
-            reminders,
-            date: '2025-08-16',
-        });
-
-        const emailArgs = sendMailMock.mock.calls[0][0];
-        expect(emailArgs.html).toContain('weekly reminders');
-    });
-
-    it('should show "reminders" when reminder types are mixed', async () => {
-        const reminders = [
-            {
-                id: 1,
-                title: 'Daily standup',
-                reminder_type: 'recurring' as const,
-                frequency: 'daily' as const,
-            },
-            {
-                id: 2,
-                title: 'Doctor appointment',
-                reminder_type: 'once' as const,
-            },
-            {
-                id: 3,
-                title: 'Weekly review',
-                reminder_type: 'recurring' as const,
-                frequency: 'weekly' as const,
-            },
-        ];
-
-        await testMailUtils.sendReminderDigestEmail({
-            email: 'test@example.com',
-            username: 'TestUser',
-            reminders,
-            date: '2025-08-16',
-        });
-
-        const emailArgs = sendMailMock.mock.calls[0][0];
-        expect(emailArgs.html).toContain('your reminders for');
-        expect(emailArgs.html).not.toContain('weekly reminders');
-        expect(emailArgs.html).not.toContain('daily reminders');
-    });
-});
-
-describe('processReminderDigests', () => {
-    const testUserId = 1;
-
-    beforeEach(async () => {
-        vi.spyOn(mailUtils, 'sendReminderDigestEmail').mockResolvedValue();
-    });
-
-    afterEach(() => {
-        vi.restoreAllMocks();
-    });
-
-    it('should find due reminders in the next 15 minutes and handle timezone correctly', async () => {
-        const now = libs.dayjs.utc();
-        const in10Minutes = now.add(10, 'minute');
-        const in20Minutes = now.add(20, 'minute');
-        const past20Minutes = now.subtract(20, 'minute');
-
-        await db('reminders').insert([
-            {
-                user_id: testUserId,
-                title: 'Due Soon',
-                reminder_type: 'once',
-                due_date: in10Minutes.toISOString(),
-            },
-            {
-                user_id: testUserId,
-                title: 'Due Later',
-                reminder_type: 'once',
-                due_date: in20Minutes.toISOString(),
-            },
-            {
-                user_id: testUserId,
-                title: 'Already Due',
-                reminder_type: 'once',
-                due_date: past20Minutes.toISOString(),
-            },
-        ]);
-
-        await mailUtils.processReminderDigests();
-
-        const remainingReminders = await db('reminders').where('title', 'Due Soon');
-        expect(remainingReminders).toHaveLength(0);
-
-        const futureReminders = await db('reminders').where('title', 'Due Later');
-        expect(futureReminders).toHaveLength(1);
-
-        const pastDueReminders = await db('reminders').where('title', 'Already Due');
-        expect(pastDueReminders).toHaveLength(1);
-    });
-
-    it('should handle recurring reminders and calculate next due date correctly', async () => {
-        const now = libs.dayjs.utc();
-        const in5Minutes = now.add(5, 'minute');
-
-        await db('reminders').insert([
-            {
-                user_id: testUserId,
-                title: 'Daily Reminder',
-                reminder_type: 'recurring',
-                frequency: 'daily',
-                due_date: in5Minutes.toISOString(),
-            },
-            {
-                user_id: testUserId,
-                title: 'Weekly Reminder',
-                reminder_type: 'recurring',
-                frequency: 'weekly',
-                due_date: in5Minutes.toISOString(),
-            },
-        ]);
-
-        await mailUtils.processReminderDigests();
-
-        const dailyReminder = await db('reminders').where('title', 'Daily Reminder').first();
-        const weeklyReminder = await db('reminders').where('title', 'Weekly Reminder').first();
-
-        expect(dailyReminder).toBeTruthy();
-        expect(weeklyReminder).toBeTruthy();
-
-        const originalDue = in5Minutes.toDate();
-        const dailyNextDue = new Date(dailyReminder.due_date);
-        const weeklyNextDue = new Date(weeklyReminder.due_date);
-
-        const originalDailyChicago = libs.dayjs.tz(originalDue, 'UTC').tz('America/Chicago');
-        const dailyNextChicago = libs.dayjs.tz(dailyNextDue, 'UTC').tz('America/Chicago');
-        expect(dailyNextChicago.hour()).toBe(originalDailyChicago.hour());
-        expect(dailyNextChicago.minute()).toBe(originalDailyChicago.minute());
-        expect(
-            dailyNextChicago.startOf('day').diff(originalDailyChicago.startOf('day'), 'day'),
-        ).toBe(1);
-
-        const weeklyDue = libs.dayjs.tz(weeklyReminder.due_date, 'UTC').tz('America/Chicago');
-        expect(weeklyDue.day()).toBe(6);
-
-        const weeklyDiff = weeklyNextDue.getTime() - originalDue.getTime();
-        expect(weeklyDiff).toBeGreaterThan(0);
-        expect(weeklyDiff).toBeLessThanOrEqual(14 * 24 * 60 * 60 * 1000);
-    });
-
-    it('should handle no due reminders gracefully', async () => {
-        const farFuture = new Date(Date.now() + 2 * 60 * 60 * 1000);
-
-        await db('reminders').insert({
-            user_id: testUserId,
-            title: 'Far Future',
-            reminder_type: 'once',
-            due_date: farFuture.toISOString(),
-        });
-
-        await expect(mailUtils.processReminderDigests()).resolves.not.toThrow();
-
-        const reminders = await db('reminders').where('title', 'Far Future');
-        expect(reminders).toHaveLength(1);
-    });
-
-    it('should use UTC for database queries and handle user timezones for email formatting', async () => {
-        const now = libs.dayjs.utc();
-        const in5Minutes = now.add(5, 'minute');
-
-        await db('reminders').insert({
-            user_id: testUserId,
-            title: 'UTC Test',
-            reminder_type: 'once',
-            due_date: in5Minutes.toISOString(),
-        });
-
-        await mailUtils.processReminderDigests();
-
-        const remainingReminders = await db('reminders').where('title', 'UTC Test');
-        expect(remainingReminders).toHaveLength(0);
-    });
-
-    it('should schedule weekly reminders for Saturday and monthly for the 1st', async () => {
-        const now = libs.dayjs.utc();
-        const in5Minutes = now.add(5, 'minute');
-
-        await db('reminders').insert([
-            {
-                user_id: testUserId,
-                title: 'Weekly Report',
-                reminder_type: 'recurring',
-                frequency: 'weekly',
-                due_date: in5Minutes.toISOString(),
-            },
-            {
-                user_id: testUserId,
-                title: 'Monthly Review',
-                reminder_type: 'recurring',
-                frequency: 'monthly',
-                due_date: in5Minutes.toISOString(),
-            },
-        ]);
-
-        await mailUtils.processReminderDigests();
-
-        const weeklyReminder = await db('reminders').where('title', 'Weekly Report').first();
-        const monthlyReminder = await db('reminders').where('title', 'Monthly Review').first();
-
-        expect(weeklyReminder).toBeTruthy();
-        expect(monthlyReminder).toBeTruthy();
-
-        const weeklyDue = libs.dayjs.tz(weeklyReminder.due_date, 'UTC').tz('America/Chicago');
-        expect(weeklyDue.day()).toBe(6);
-
-        const monthlyDue = libs.dayjs.tz(monthlyReminder.due_date, 'UTC').tz('America/Chicago');
-        expect(monthlyDue.date()).toBe(1);
-    });
-
-    it('should allow daily reminders to be processed multiple times', async () => {
-        const now = libs.dayjs.utc();
-        const in5Minutes = now.add(5, 'minute');
-
-        await db('reminders').insert({
-            user_id: testUserId,
-            title: 'Multi-Day Daily Reminder',
-            reminder_type: 'recurring',
-            frequency: 'daily',
-            due_date: in5Minutes.toISOString(),
-        });
-
-        await mailUtils.processReminderDigests();
-
-        let reminder = await db('reminders').where('title', 'Multi-Day Daily Reminder').first();
-        expect(reminder).toBeTruthy();
-
-        const firstDue = new Date(reminder.due_date);
-        expect(firstDue.getTime()).toBeGreaterThan(in5Minutes.toDate().getTime());
-
-        const nowAgain = libs.dayjs.utc();
-        const in5MinutesAgain = nowAgain.add(5, 'minute');
-
-        await db('reminders')
-            .where('title', 'Multi-Day Daily Reminder')
-            .update({ due_date: in5MinutesAgain.toISOString() });
-
-        await mailUtils.processReminderDigests();
-
-        reminder = await db('reminders').where('title', 'Multi-Day Daily Reminder').first();
-        expect(reminder).toBeTruthy();
-
-        const originalChicago = libs.dayjs
-            .tz(in5MinutesAgain.toDate(), 'UTC')
-            .tz('America/Chicago');
-        const updatedChicago = libs.dayjs.tz(reminder.due_date, 'UTC').tz('America/Chicago');
-        expect(updatedChicago.hour()).toBe(originalChicago.hour());
-        expect(updatedChicago.minute()).toBe(originalChicago.minute());
-        expect(updatedChicago.startOf('day').diff(originalChicago.startOf('day'), 'day')).toBe(1);
     });
 });
